Add support for ymodem and TI Serial flash loader
The `pyTISerialFlash.py` contains a `FlashLdr` class similar
to existing flash loader classes. There are memory functions
that are not applicable because the ymodem protocol is used
instead of character by character and page by page methods.

Also a new method `download_application` is new in the
`FlashLdr` class to download the actual file.

The `YmodemMCU` class in the `pyYmodem.py` file is a thin
wrapper around the ymodem library. It also has an extensive
`main()` function that was used to check application file
upload and download functionality together with a terminal
emulator (minicom).