Installation for ymodem support
The `modem` package used for ymodem functionality is getting
old and was originally written for Python2. A fork was found
that had ported it to Python 3. As some issues were found and
fixed in the ymodem protocol, this repo was forked, fixed and
a PR issued upstream.

Until these PR's get merged and the updates published to `pypi.org`
this wheel file can be used.